+++ src/main/kotlin/com/gitlab/eclipse/preferences/healthcheck/FeatureStateParams.kt
@@ -1,0 +1,6 @@
+package com.gitlab.eclipse.preferences.healthcheck
+
+data class FeatureStateParams(
+  val featureId: String,
+  val engagedChecks: List<FeatureStateCheckParams>
+)